backend/requirements.txt
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
pydantic[email]
python-multipart
pydantic-settings
openai
PyMuPDF
requests
bcrypt==4.0.1
google-generativeai
